+++ packages/react/src/primitives/index.ts
@@ -17,3 +17,4 @@ export { useMessagePartFile } from "./messagePart/useMessagePartFile";
 export { useMessagePartImage } from "./messagePart/useMessagePartImage";
 export { useMessagePartData } from "./messagePart/useMessagePartData";
 export { useThreadViewportAutoScroll } from "./thread/useThreadViewportAutoScroll";
+export { useScrollLock } from "./reasoning";