+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,23 @@
 # assistant-ui
 
+## 0.0.64
+
+### Patch Changes
+
+- 2fc7e99: chore: update deps
+
+## 0.0.63
+
+### Patch Changes
+
+- 953db24: chore: update deps
+
+## 0.0.62
+
+### Patch Changes
+
+- chore: update deps
+
 ## 0.0.61
 
 ### Patch Changes