+++ packages/create-assistant-ui/CHANGELOG.md
@@ -1,5 +1,23 @@
 # create-assistant-ui
 
+## 0.0.31
+
+### Patch Changes
+
+- 2fc7e99: chore: update deps
+
+## 0.0.30
+
+### Patch Changes
+
+- 953db24: chore: update deps
+
+## 0.0.29
+
+### Patch Changes
+
+- chore: update deps
+
 ## 0.0.28
 
 ### Patch Changes